// src/payload-types.ts
export interface Media {
    id: number;
    alt: string;
    updatedAt: string;
    createdAt: string;
    url?: string | null;
    thumbnailURL?: string | null;
    filename?: string | null;
    mimeType?: string | null;
    filesize?: number | null;
    width?: number | null;
    height?: number | null;
    focalX?: number | null;
    focalY?: number | null;
}
// ...
export interface Homepage {
    id: number;
    heroImage: number | Media;
    heroTitle: string;
    heroSubtitle: string;
    ctaLabel: string;
    sensorHubTitle: string;
    sensorHubDescription: string;
    sensorHubVideo: number | Media;
    recordingsTitle: string;
    recordingsDescription: string;
    recordingDecorVideo: number | Media;
    readingsTitle: string;
    readingsDescription: string;
    aboutSectionTitle: string;
    aboutSpotifyEmbedUrl: string;
    gallery?:
        | {
              image: number | Media;
              alt?: string | null;
              id?: string | null;
          }[]
        | null;
    updatedAt?: string | null;
    createdAt?: string | null;
}